Split Toggles into individual folders
