+++ src/components/DetailsOverlay/Toggles/ExpandableRowsToggle/index.tsx
@@ -1,0 +1,20 @@
+import { QueryParams } from "constants/queryParams";
+import { useQueryParam } from "hooks/useQueryParam";
+import BaseToggle from "../BaseToggle";
+
+const ExpandableRowsToggle: React.FC = () => {
+  const [expandableRows, setExpandableRows] = useQueryParam(
+    QueryParams.Expandable,
+    false
+  );
+  return (
+    <BaseToggle
+      data-cy="expandable-rows-toggle"
+      label="Expandable Rows"
+      onChange={setExpandableRows}
+      value={expandableRows}
+    />
+  );
+};
+
+export default ExpandableRowsToggle;
