+++ src/components/DetailsOverlay/Toggles/index.tsx
@@ -1,0 +1,15 @@
+import BaseToggle from "./BaseToggle";
+import CaseSensitiveToggle from "./CaseSensitiveToggle";
+import ExpandableRowsToggle from "./ExpandableRowsToggle";
+import FilterLogicToggle from "./FilterLogicToggle";
+import PrettyPrintToggle from "./PrettyPrintToggle";
+import WrapToggle from "./WrapToggle";
+
+export {
+  BaseToggle,
+  CaseSensitiveToggle,
+  ExpandableRowsToggle,
+  FilterLogicToggle,
+  PrettyPrintToggle,
+  WrapToggle,
+};
